Moved port definitions into a dedicated config file
