+++ src/server/gameServer.ts
@@ -3,7 +3,7 @@ import { Server } from 'boardgame.io/server';
 import { v4 as uuidv4 } from 'uuid';
 
 import { ElevationOfPrivilege } from '../game/eop';
-import { INTERNAL_API_PORT, SERVER_PORT } from '../utils/constants';
+import { INTERNAL_API_PORT, SERVER_PORT } from '../utils/serverConfig';
 import { getDatabase } from './config';
 
 import type { ModelFlatFile } from './ModelFlatFile';
